Convert the exist core functionality to compile using maven.
The main src folder contains the core of exist.  In order to get this
to compile, the necessary poms needed to be found.  This initial
implementation just gets things to compile, all dependencies come
from maven central.

It seems Jetty 5 was used as the basis for existdb, so it should be updated
to jetty 7/8 at some point.

Future commits will break out the extensions into their own modules.  Ultimate
goal is to contribute this back to the eXist-db project.