More Test clean up.
Many of the tests leave a dirty database after they run.  This affects other
subsequent tests.  To address this an AbstractDBTest class was created and
a cleanUp method as placed there.  All tests where possible should extend
from AbstractDBTest and in the setup method make sure they call cleanUp.
This will guaranteed a clean database for the tests.

Change-Id: I4f89fe1130461c5bd61ce1225799951d0b0e57d9